@@ -25,20 +25,21 @@
         "@types/glob": "^7.2.0",
         "@types/node": "^17.0.45",
         "@types/yargs": "^16.0.5",
-        "@typescript-eslint/eslint-plugin": "^5.59.5",
-        "@typescript-eslint/parser": "^5.59.5",
+        "@typescript-eslint/eslint-plugin": "^6.1.0",
+        "@typescript-eslint/parser": "^6.1.0",
         "detect-indent": "^6.1.0",
-        "eslint": "^8.40.0",
+        "eslint": "^8.45.0",
         "eslint-config-prettier": "^8.8.0",
-        "eslint-plugin-simple-import-sort": "^7.0.0",
+        "eslint-plugin-simple-import-sort": "^10.0.0",
         "glob": "^7.2.3",
         "jsonc-parser": "^3.2.0",
         "lerna": "^6.6.2",
-        "npm-check-updates": "^16.10.12",
-        "p-queue": "^6.6.2",
+        "npm-check-updates": "^16.10.16",
+        "p-queue": "^7.3.4",
         "prettier": "2.8.8",
-        "syncpack": "^9.8.6",
+        "syncpack": "^10.7.3",
         "typescript": "~4.4.4",
+        "word-wrap": "1.2.4",
         "yargs": "^16.2.0"
     }
 }
