@@ -2,7 +2,7 @@
     "name": "@replit/pyright-extended",
     "displayName": "pyright-extended",
     "description": "Extending pyright with yapf + ruff",
-    "version": "2.0.1",
+    "version": "2.0.2",
     "license": "MIT",
     "author": {
         "name": "Replit"
@@ -27,16 +27,19 @@
         "prepack": "npm run clean && shx cp ../../README.md . && shx cp ../../LICENSE.txt . && npm run build",
         "webpack": "webpack --mode development --progress"
     },
+    "optionalDependencies": {
+        "fsevents": "~2.3.2"
+    },
     "devDependencies": {
         "@types/copy-webpack-plugin": "^10.1.0",
         "@types/node": "^17.0.45",
         "copy-webpack-plugin": "^11.0.0",
         "esbuild-loader": "^3.0.1",
         "shx": "^0.3.4",
-        "ts-loader": "^9.4.3",
+        "ts-loader": "^9.4.4",
         "typescript": "~4.4.4",
-        "webpack": "^5.85.0",
-        "webpack-cli": "^5.1.1"
+        "webpack": "^5.88.1",
+        "webpack-cli": "^5.1.4"
     },
     "files": [
         "/dist",
