@@ -2,7 +2,7 @@
     "name": "vscode-pyright",
     "displayName": "Pyright",
     "description": "VS Code static type checking for Python",
-    "version": "1.1.317",
+    "version": "1.1.318",
     "private": true,
     "license": "MIT",
     "author": {
@@ -951,7 +951,7 @@
     },
     "dependencies": {
         "vscode-jsonrpc": "8.1.0",
-        "vscode-languageclient": "^8.1.0",
+        "vscode-languageclient": "8.1.0",
         "vscode-languageserver": "8.1.0",
         "vscode-languageserver-protocol": "3.17.3"
     },
@@ -963,10 +963,10 @@
         "detect-indent": "^6.1.0",
         "esbuild-loader": "^3.0.1",
         "shx": "^0.3.4",
-        "ts-loader": "^9.4.3",
+        "ts-loader": "^9.4.4",
         "typescript": "~4.4.4",
         "vsce": "^2.7.0",
-        "webpack": "^5.85.0",
-        "webpack-cli": "^5.1.1"
+        "webpack": "^5.88.1",
+        "webpack-cli": "^5.1.4"
     }
 }
