@@ -31,7 +31,7 @@ import {
 import { getFileInfo } from './analyzerNodeInfo';
 import { CodeFlowReferenceExpressionNode } from './codeFlowTypes';
 import { populateTypeVarContextBasedOnExpectedType } from './constraintSolver';
-import { isMatchingExpression } from './parseTreeUtils';
+import { getTypeVarScopesForNode, isMatchingExpression } from './parseTreeUtils';
 import { getTypedDictMembersForClass } from './typedDicts';
 import { EvaluatorFlags, TypeEvaluator, TypeResult } from './typeEvaluatorTypes';
 import {
@@ -61,6 +61,7 @@ import {
 import {
     addConditionToType,
     applySolvedTypeVars,
+    containsAnyOrUnknown,
     convertToInstance,
     doForEachSubtype,
     getTypeCondition,
@@ -97,11 +98,11 @@ const classPatternSpecialCases = [
 
 interface SequencePatternInfo {
     subtype: Type;
-    definiteNoMatch: boolean;
+    isDefiniteNoMatch: boolean;
+    isPotentialNoMatch?: boolean;
     entryTypes: Type[];
     isIndeterminateLength?: boolean;
     isTuple?: boolean;
-    isObject?: boolean;
 }
 
 interface MappingPatternInfo {
@@ -194,11 +195,11 @@ function narrowTypeBasedOnSequencePattern(
     isPositiveTest: boolean
 ): Type {
     type = transformPossibleRecursiveTypeAlias(type);
-    let sequenceInfo = getSequencePatternInfo(evaluator, type, pattern.entries.length, pattern.starEntryIndex);
+    let sequenceInfo = getSequencePatternInfo(evaluator, pattern, type);
 
     // Further narrow based on pattern entry types.
     sequenceInfo = sequenceInfo.filter((entry) => {
-        if (entry.definiteNoMatch) {
+        if (entry.isDefiniteNoMatch) {
             if (isPositiveTest) {
                 return false;
             } else {
@@ -235,8 +236,7 @@ function narrowTypeBasedOnSequencePattern(
                 index,
                 pattern.entries.length,
                 pattern.starEntryIndex,
-                /* unpackStarEntry */ true,
-                /* isSubjectObject */ false
+                /* unpackStarEntry */ true
             );
 
             const narrowedEntryType = narrowTypeBasedOnPattern(evaluator, entryType, sequenceEntry, isPositiveTest);
@@ -265,6 +265,10 @@ function narrowTypeBasedOnSequencePattern(
                     }
                 }
             } else {
+                if (entry.isPotentialNoMatch) {
+                    isDefiniteMatch = false;
+                }
+
                 if (!isNever(narrowedEntryType)) {
                     isDefiniteMatch = false;
 
@@ -326,16 +330,18 @@ function narrowTypeBasedOnSequencePattern(
                 }
             }
 
-            // If this is an object, we can narrow it to a specific Sequence type.
-            if (entry.isObject) {
+            // If this is a supertype of Sequence, we can narrow it to a Sequence type.
+            if (entry.isPotentialNoMatch) {
                 const sequenceType = evaluator.getTypingType(pattern, 'Sequence');
                 if (sequenceType && isInstantiableClass(sequenceType)) {
+                    let typeArgType = evaluator.stripLiteralValue(combineTypes(narrowedEntryTypes));
+
+                    // If the type is a union that contains Any or Unknown, remove the other types
+                    // before wrapping it in a Sequence.
+                    typeArgType = containsAnyOrUnknown(typeArgType, /* recurse */ false) ?? typeArgType;
+
                     entry.subtype = ClassType.cloneAsInstance(
-                        ClassType.cloneForSpecialization(
-                            sequenceType,
-                            [evaluator.stripLiteralValue(combineTypes(narrowedEntryTypes))],
-                            /* isTypeArgumentExplicit */ true
-                        )
+                        ClassType.cloneForSpecialization(sequenceType, [typeArgType], /* isTypeArgumentExplicit */ true)
                     );
                 }
             }
@@ -1098,10 +1104,11 @@ function getMappingPatternInfo(evaluator: TypeEvaluator, type: Type): MappingPat
 // sufficient length, it sets definiteNoMatch to true.
 function getSequencePatternInfo(
     evaluator: TypeEvaluator,
-    type: Type,
-    entryCount: number,
-    starEntryIndex: number | undefined
+    pattern: PatternSequenceNode,
+    type: Type
 ): SequencePatternInfo[] {
+    const entryCount = pattern.entries.length;
+    const starEntryIndex = pattern.starEntryIndex;
     const sequenceInfo: SequencePatternInfo[] = [];
     const minEntryCount = starEntryIndex === undefined ? entryCount : entryCount - 1;
 
@@ -1110,28 +1117,7 @@ function getSequencePatternInfo(
         let mroClassToSpecialize: ClassType | undefined;
         let pushedEntry = false;
 
-        if (isAnyOrUnknown(concreteSubtype)) {
-            sequenceInfo.push({
-                subtype,
-                entryTypes: [concreteSubtype],
-                isIndeterminateLength: true,
-                definiteNoMatch: false,
-            });
-            return;
-        }
-
         if (isClassInstance(concreteSubtype)) {
-            if (ClassType.isBuiltIn(concreteSubtype, 'object')) {
-                sequenceInfo.push({
-                    subtype,
-                    entryTypes: [convertToInstance(concreteSubtype)],
-                    isIndeterminateLength: true,
-                    isObject: true,
-                    definiteNoMatch: false,
-                });
-                return;
-            }
-
             for (const mroClass of concreteSubtype.details.mro) {
                 if (!isInstantiableClass(mroClass)) {
                     break;
@@ -1143,7 +1129,7 @@ function getSequencePatternInfo(
                     ClassType.isBuiltIn(mroClass, 'bytes') ||
                     ClassType.isBuiltIn(mroClass, 'bytearray')
                 ) {
-                    break;
+                    return;
                 }
 
                 if (ClassType.isBuiltIn(mroClass, 'Sequence')) {
@@ -1168,7 +1154,7 @@ function getSequencePatternInfo(
                                 entryTypes: [combineTypes(specializedSequence.tupleTypeArguments.map((t) => t.type))],
                                 isIndeterminateLength: true,
                                 isTuple: true,
-                                definiteNoMatch: false,
+                                isDefiniteNoMatch: false,
                             });
                             pushedEntry = true;
                         } else {
@@ -1182,7 +1168,7 @@ function getSequencePatternInfo(
                                     entryTypes: specializedSequence.tupleTypeArguments.map((t) => t.type),
                                     isIndeterminateLength: false,
                                     isTuple: true,
-                                    definiteNoMatch: false,
+                                    isDefiniteNoMatch: false,
                                 });
                                 pushedEntry = true;
                             }
@@ -1197,20 +1183,73 @@ function getSequencePatternInfo(
                                 : UnknownType.create(),
                         ],
                         isIndeterminateLength: true,
-                        definiteNoMatch: false,
+                        isDefiniteNoMatch: false,
                     });
                     pushedEntry = true;
                 }
             }
         }
 
-        // Push an entry that indicates that this is definitely not a match.
         if (!pushedEntry) {
+            // If it wasn't a subtype of Sequence, see if it's a supertype.
+            const sequenceType = evaluator.getTypingType(pattern, 'Sequence');
+
+            if (sequenceType && isInstantiableClass(sequenceType)) {
+                const sequenceTypeVarContext = new TypeVarContext(getTypeVarScopeId(sequenceType));
+                if (
+                    populateTypeVarContextBasedOnExpectedType(
+                        evaluator,
+                        ClassType.cloneAsInstance(sequenceType),
+                        subtype,
+                        sequenceTypeVarContext,
+                        getTypeVarScopesForNode(pattern),
+                        pattern.start
+                    )
+                ) {
+                    const specializedSequence = applySolvedTypeVars(
+                        ClassType.cloneAsInstantiable(sequenceType),
+                        sequenceTypeVarContext
+                    ) as ClassType;
+
+                    if (specializedSequence.typeArguments && specializedSequence.typeArguments.length > 0) {
+                        sequenceInfo.push({
+                            subtype,
+                            entryTypes: [specializedSequence.typeArguments[0]],
+                            isIndeterminateLength: true,
+                            isDefiniteNoMatch: false,
+                            isPotentialNoMatch: true,
+                        });
+                        return;
+                    }
+                }
+
+                if (
+                    evaluator.assignType(
+                        subtype,
+                        ClassType.cloneForSpecialization(
+                            ClassType.cloneAsInstance(sequenceType),
+                            [UnknownType.create()],
+                            /* isTypeArgumentExplicit */ true
+                        )
+                    )
+                ) {
+                    sequenceInfo.push({
+                        subtype,
+                        entryTypes: [UnknownType.create()],
+                        isIndeterminateLength: true,
+                        isDefiniteNoMatch: false,
+                        isPotentialNoMatch: true,
+                    });
+                    return;
+                }
+            }
+
+            // Push an entry that indicates that this is definitely not a match.
             sequenceInfo.push({
                 subtype,
                 entryTypes: [],
                 isIndeterminateLength: true,
-                definiteNoMatch: true,
+                isDefiniteNoMatch: true,
             });
         }
     });
@@ -1225,21 +1264,10 @@ function getTypeOfPatternSequenceEntry(
     entryIndex: number,
     entryCount: number,
     starEntryIndex: number | undefined,
-    unpackStarEntry: boolean,
-    isSubjectObject: boolean
+    unpackStarEntry: boolean
 ): Type {
     if (sequenceInfo.isIndeterminateLength) {
         let entryType = sequenceInfo.entryTypes[0];
-
-        // If the subject is typed as an "object", then the star entry
-        // is simply a list[object]. Without this special case, the list
-        // will be typed based on the union of all elements in the sequence.
-        if (isSubjectObject) {
-            const objectType = evaluator.getBuiltInObject(node, 'object');
-            if (objectType && isClassInstance(objectType)) {
-                entryType = objectType;
-            }
-        }
 
         if (!unpackStarEntry && entryIndex === starEntryIndex && !isNever(entryType)) {
             entryType = wrapTypeInList(evaluator, node, entryType);
@@ -1277,25 +1305,21 @@ function getTypeOfPatternSequenceEntry(
 }
 
 // Recursively assigns the specified type to the pattern and any capture
-// nodes within it
+// nodes within it. It returns the narrowed type, as dictated by the pattern.
 export function assignTypeToPatternTargets(
     evaluator: TypeEvaluator,
     type: Type,
     isTypeIncomplete: boolean,
-    isSubjectObject: boolean,
     pattern: PatternAtomNode
-) {
+): Type {
     // Further narrow the type based on this pattern.
-    type = narrowTypeBasedOnPattern(evaluator, type, pattern, /* positiveTest */ true);
+    const narrowedType = narrowTypeBasedOnPattern(evaluator, type, pattern, /* positiveTest */ true);
 
     switch (pattern.nodeType) {
         case ParseNodeType.PatternSequence: {
-            const sequenceInfo = getSequencePatternInfo(
-                evaluator,
-                type,
-                pattern.entries.length,
-                pattern.starEntryIndex
-            ).filter((seqInfo) => !seqInfo.definiteNoMatch);
+            const sequenceInfo = getSequencePatternInfo(evaluator, pattern, narrowedType).filter(
+                (seqInfo) => !seqInfo.isDefiniteNoMatch
+            );
 
             pattern.entries.forEach((entry, index) => {
                 const entryType = combineTypes(
@@ -1307,28 +1331,33 @@ export function assignTypeToPatternTargets(
                             index,
                             pattern.entries.length,
                             pattern.starEntryIndex,
-                            /* unpackStarEntry */ false,
-                            isSubjectObject
+                            /* unpackStarEntry */ false
                         )
                     )
                 );
 
-                assignTypeToPatternTargets(evaluator, entryType, isTypeIncomplete, /* isSubjectObject */ false, entry);
+                assignTypeToPatternTargets(evaluator, entryType, isTypeIncomplete, entry);
             });
             break;
         }
 
         case ParseNodeType.PatternAs: {
             if (pattern.target) {
-                evaluator.assignTypeToExpression(pattern.target, type, isTypeIncomplete, pattern.target);
+                evaluator.assignTypeToExpression(pattern.target, narrowedType, isTypeIncomplete, pattern.target);
             }
 
+            let runningNarrowedType = narrowedType;
             pattern.orPatterns.forEach((orPattern) => {
-                assignTypeToPatternTargets(evaluator, type, isTypeIncomplete, isSubjectObject, orPattern);
+                assignTypeToPatternTargets(evaluator, runningNarrowedType, isTypeIncomplete, orPattern);
 
                 // OR patterns are evaluated left to right, so we can narrow
                 // the type as we go.
-                type = narrowTypeBasedOnPattern(evaluator, type, orPattern, /* positiveTest */ false);
+                runningNarrowedType = narrowTypeBasedOnPattern(
+                    evaluator,
+                    runningNarrowedType,
+                    orPattern,
+                    /* positiveTest */ false
+                );
             });
             break;
         }
@@ -1337,19 +1366,19 @@ export function assignTypeToPatternTargets(
             if (pattern.isWildcard) {
                 if (!isTypeIncomplete) {
                     const fileInfo = getFileInfo(pattern);
-                    if (isUnknown(type)) {
+                    if (isUnknown(narrowedType)) {
                         evaluator.addDiagnostic(
                             fileInfo.diagnosticRuleSet.reportUnknownVariableType,
                             DiagnosticRule.reportUnknownVariableType,
                             Localizer.Diagnostic.wildcardPatternTypeUnknown(),
                             pattern.target
                         );
-                    } else if (isPartlyUnknown(type)) {
+                    } else if (isPartlyUnknown(narrowedType)) {
                         const diagAddendum = new DiagnosticAddendum();
                         diagAddendum.addMessage(
                             Localizer.DiagnosticAddendum.typeOfSymbol().format({
                                 name: '_',
-                                type: evaluator.printType(type, { expandTypeAlias: true }),
+                                type: evaluator.printType(narrowedType, { expandTypeAlias: true }),
                             })
                         );
                         evaluator.addDiagnostic(
@@ -1361,13 +1390,13 @@ export function assignTypeToPatternTargets(
                     }
                 }
             } else {
-                evaluator.assignTypeToExpression(pattern.target, type, isTypeIncomplete, pattern.target);
+                evaluator.assignTypeToExpression(pattern.target, narrowedType, isTypeIncomplete, pattern.target);
             }
             break;
         }
 
         case ParseNodeType.PatternMapping: {
-            const mappingInfo = getMappingPatternInfo(evaluator, type);
+            const mappingInfo = getMappingPatternInfo(evaluator, narrowedType);
 
             pattern.entries.forEach((mappingEntry) => {
                 const keyTypes: Type[] = [];
@@ -1432,20 +1461,8 @@ export function assignTypeToPatternTargets(
                 const valueType = combineTypes(valueTypes);
 
                 if (mappingEntry.nodeType === ParseNodeType.PatternMappingKeyEntry) {
-                    assignTypeToPatternTargets(
-                        evaluator,
-                        keyType,
-                        isTypeIncomplete,
-                        /* isSubjectObject */ false,
-                        mappingEntry.keyPattern
-                    );
-                    assignTypeToPatternTargets(
-                        evaluator,
-                        valueType,
-                        isTypeIncomplete,
-                        /* isSubjectObject */ false,
-                        mappingEntry.valuePattern
-                    );
+                    assignTypeToPatternTargets(evaluator, keyType, isTypeIncomplete, mappingEntry.keyPattern);
+                    assignTypeToPatternTargets(evaluator, valueType, isTypeIncomplete, mappingEntry.valuePattern);
                 } else if (mappingEntry.nodeType === ParseNodeType.PatternMappingExpandEntry) {
                     const dictClass = evaluator.getBuiltInType(pattern, 'dict');
                     const strType = evaluator.getBuiltInObject(pattern, 'str');
@@ -1473,9 +1490,9 @@ export function assignTypeToPatternTargets(
         case ParseNodeType.PatternClass: {
             const argTypes: Type[][] = pattern.arguments.map((arg) => []);
 
-            evaluator.mapSubtypesExpandTypeVars(type, /* conditionFilter */ undefined, (expandedSubtype) => {
+            evaluator.mapSubtypesExpandTypeVars(narrowedType, /* conditionFilter */ undefined, (expandedSubtype) => {
                 if (isClassInstance(expandedSubtype)) {
-                    doForEachSubtype(type, (subjectSubtype) => {
+                    doForEachSubtype(narrowedType, (subjectSubtype) => {
                         const concreteSubtype = evaluator.makeTopLevelTypeVarsConcrete(subjectSubtype);
 
                         if (isAnyOrUnknown(concreteSubtype)) {
@@ -1516,13 +1533,7 @@ export function assignTypeToPatternTargets(
             });
 
             pattern.arguments.forEach((arg, index) => {
-                assignTypeToPatternTargets(
-                    evaluator,
-                    combineTypes(argTypes[index]),
-                    isTypeIncomplete,
-                    /* isSubjectObject */ false,
-                    arg.pattern
-                );
+                assignTypeToPatternTargets(evaluator, combineTypes(argTypes[index]), isTypeIncomplete, arg.pattern);
             });
             break;
         }
@@ -1534,6 +1545,8 @@ export function assignTypeToPatternTargets(
             break;
         }
     }
+
+    return narrowedType;
 }
 
 function wrapTypeInList(evaluator: TypeEvaluator, node: ParseNode, type: Type): Type {
@@ -1543,6 +1556,10 @@ function wrapTypeInList(evaluator: TypeEvaluator, node: ParseNode, type: Type): 
 
     const listObjectType = convertToInstance(evaluator.getBuiltInObject(node, 'list'));
     if (listObjectType && isClassInstance(listObjectType)) {
+        // If the type is a union that contains an Any or Unknown, eliminate the other
+        // types before wrapping it in a list.
+        type = containsAnyOrUnknown(type, /* recurse */ false) ?? type;
+
         return ClassType.cloneForSpecialization(listObjectType, [type], /* isTypeArgumentExplicit */ true);
     }
 
