@@ -222,12 +222,12 @@ export function getTypeNarrowingCallback(
                 }
             }
 
-            // Look for "X is Y" or "X is not Y" where Y is a an enum or bool literal.
             if (isOrIsNotOperator) {
                 if (ParseTreeUtils.isMatchingExpression(reference, testExpression.leftExpression)) {
                     const rightTypeResult = evaluator.getTypeOfExpression(testExpression.rightExpression);
                     const rightType = rightTypeResult.type;
 
+                    // Look for "X is Y" or "X is not Y" where Y is a an enum or bool literal.
                     if (
                         isClassInstance(rightType) &&
                         (ClassType.isEnumClass(rightType) || ClassType.isBuiltIn(rightType, 'bool')) &&
@@ -246,9 +246,19 @@ export function getTypeNarrowingCallback(
                             };
                         };
                     }
+
+                    // Look for X is <class> or X is not <class>.
+                    if (isInstantiableClass(rightType)) {
+                        return (type: Type) => {
+                            return {
+                                type: narrowTypeForClassComparison(evaluator, type, rightType, adjIsPositiveTest),
+                                isIncomplete: !!rightTypeResult.isIncomplete,
+                            };
+                        };
+                    }
                 }
 
-                // Look for X[<literal>] is <literal> or X[<literal>] is not <literal>
+                // Look for X[<literal>] is <literal> or X[<literal>] is not <literal>.
                 if (
                     testExpression.leftExpression.nodeType === ParseNodeType.Index &&
                     testExpression.leftExpression.items.length === 1 &&
@@ -997,7 +1007,9 @@ function narrowTypeForIsNone(evaluator: TypeEvaluator, type: Type, isPositiveTes
         return transformPossibleRecursiveTypeAlias(subtype);
     });
 
-    return evaluator.mapSubtypesExpandTypeVars(
+    let resultIncludesNoneSubtype = false;
+
+    const result = evaluator.mapSubtypesExpandTypeVars(
         expandedType,
         /* conditionFilter */ undefined,
         (subtype, unexpandedSubtype) => {
@@ -1017,6 +1029,7 @@ function narrowTypeForIsNone(evaluator: TypeEvaluator, type: Type, isPositiveTes
 
             // See if it's a match for object.
             if (isClassInstance(subtype) && ClassType.isBuiltIn(subtype, 'object')) {
+                resultIncludesNoneSubtype = true;
                 return isPositiveTest
                     ? addConditionToType(NoneType.createInstance(), subtype.condition)
                     : adjustedSubtype;
@@ -1024,12 +1037,24 @@ function narrowTypeForIsNone(evaluator: TypeEvaluator, type: Type, isPositiveTes
 
             // See if it's a match for None.
             if (isNoneInstance(subtype) === isPositiveTest) {
+                resultIncludesNoneSubtype = true;
                 return subtype;
             }
 
             return undefined;
         }
     );
+
+    // If this is a positive test and the result is a union that includes None,
+    // we can eliminate all the non-None subtypes include Any or Unknown. If some
+    // of the subtypes are None types with conditions, retain those.
+    if (isPositiveTest && resultIncludesNoneSubtype) {
+        return mapSubtypes(result, (subtype) => {
+            return isNoneInstance(subtype) ? subtype : undefined;
+        });
+    }
+
+    return result;
 }
 
 // Handle type narrowing for expressions of the form "x is ..." and "x is not ...".
@@ -1801,11 +1826,7 @@ function narrowTypeForTypedDictKey(
 
             if (isPositiveTest) {
                 if (!tdEntry) {
-                    // If the class is final, we can say with certainty that if
-                    // the TypedDict doesn't define this entry, it is not this type.
-                    // If it's not final, we can't say this because it could be a
-                    // subclass of this TypedDict that adds more fields.
-                    return ClassType.isFinal(subtype) ? undefined : subtype;
+                    return undefined;
                 }
 
                 // If the entry is currently not required and not marked provided, we can mark
@@ -2065,6 +2086,54 @@ function narrowTypeForTypeIs(evaluator: TypeEvaluator, type: Type, classType: Cl
             return unexpandedSubtype;
         }
     );
+}
+
+// Attempts to narrow a type based on a comparison with a class using "is" or
+// "is not". This pattern is sometimes used for sentinels.
+function narrowTypeForClassComparison(
+    evaluator: TypeEvaluator,
+    referenceType: Type,
+    classType: ClassType,
+    isPositiveTest: boolean
+): Type {
+    return mapSubtypes(referenceType, (subtype) => {
+        const concreteSubtype = evaluator.makeTopLevelTypeVarsConcrete(subtype);
+
+        if (isPositiveTest) {
+            if (isNoneInstance(concreteSubtype)) {
+                return undefined;
+            }
+
+            if (isClassInstance(concreteSubtype) && TypeBase.isInstance(subtype)) {
+                return undefined;
+            }
+
+            if (isInstantiableClass(concreteSubtype) && ClassType.isFinal(concreteSubtype)) {
+                if (
+                    !ClassType.isSameGenericClass(concreteSubtype, classType) &&
+                    !isIsinstanceFilterSuperclass(
+                        evaluator,
+                        concreteSubtype,
+                        classType,
+                        classType,
+                        /* isInstanceCheck */ false
+                    )
+                ) {
+                    return undefined;
+                }
+            }
+        } else {
+            if (
+                isInstantiableClass(concreteSubtype) &&
+                ClassType.isSameGenericClass(classType, concreteSubtype) &&
+                ClassType.isFinal(classType)
+            ) {
+                return undefined;
+            }
+        }
+
+        return subtype;
+    });
 }
 
 // Attempts to narrow a type (make it more constrained) based on a comparison
