@@ -734,7 +734,11 @@ export namespace ClassType {
             return type.cached.typeBaseInstantiableType as ClassType;
         }
 
-        return TypeBase.cloneTypeAsInstantiable(type, /* cache */ true);
+        const newInstance = TypeBase.cloneTypeAsInstantiable(type, /* cache */ true);
+        newInstance.flags &= ~TypeFlags.SpecialForm;
+        newInstance.includeSubclasses = true;
+
+        return newInstance;
     }
 
     export function cloneForSpecialization(
@@ -2515,6 +2519,12 @@ export namespace TypeVarType {
 
         // By this point, the variance should have been inferred.
         assert(variance !== Variance.Auto, 'Expected variance to be inferred');
+
+        // If we're in the process of computing variance, it will still be
+        // unknown. Default to covariant in this case.
+        if (variance === Variance.Unknown) {
+            return Variance.Covariant;
+        }
 
         return variance;
     }
