@@ -1,30 +1,22 @@
 # This sample tests type narrowing for TypedDict types based
 # on whether a key is in or not in the dict.
 
-from typing import TypedDict, final
+from typing import TypedDict
 
 
-@final
 class TD1(TypedDict):
     a: str
     b: int
 
 
-@final
 class TD2(TypedDict):
     a: int
     c: str
 
 
-@final
 class TD3(TypedDict, total=False):
     a: int
     d: str
-
-
-class TD4(TypedDict):
-    a: int
-    c: str
 
 
 def f1(p: TD1 | TD2):
@@ -91,10 +83,3 @@ def f7(p: TD3):
 def f8(p: TD3):
     if "a" in p:
         f7(p)
-
-
-def f9(p: TD1 | TD4):
-    if "b" in p:
-        reveal_type(p, expected_text="TD1 | TD4")
-    else:
-        reveal_type(p, expected_text="TD4")
