@@ -19,11 +19,11 @@ class Charset:
     def get_body_encoding(self) -> str | Callable[[Message], None]: ...
     def get_output_charset(self) -> str | None: ...
     def header_encode(self, string: str) -> str: ...
-    def header_encode_lines(self, string: str, maxlengths: Iterator[int]) -> list[str]: ...
+    def header_encode_lines(self, string: str, maxlengths: Iterator[int]) -> list[str | None]: ...
     @overload
     def body_encode(self, string: None) -> None: ...
     @overload
-    def body_encode(self, string: str) -> str: ...
+    def body_encode(self, string: str | bytes) -> str: ...
     def __eq__(self, other: object) -> bool: ...
     def __ne__(self, __value: object) -> bool: ...
 
