@@ -53,7 +53,7 @@ compat32: Compat32
 class EmailPolicy(Policy):
     utf8: bool
     refold_source: str
-    header_factory: Callable[[str, str], str]
+    header_factory: Callable[[str, Any], Any]
     content_manager: ContentManager
     def __init__(
         self,
@@ -70,9 +70,9 @@ class EmailPolicy(Policy):
         content_manager: ContentManager = ...,
     ) -> None: ...
     def header_source_parse(self, sourcelines: list[str]) -> tuple[str, str]: ...
-    def header_store_parse(self, name: str, value: str) -> tuple[str, str]: ...
-    def header_fetch_parse(self, name: str, value: str) -> str: ...
-    def fold(self, name: str, value: str) -> str: ...
+    def header_store_parse(self, name: str, value: Any) -> tuple[str, Any]: ...
+    def header_fetch_parse(self, name: str, value: str) -> Any: ...
+    def fold(self, name: str, value: str) -> Any: ...
     def fold_binary(self, name: str, value: str) -> bytes: ...
 
 default: EmailPolicy
