@@ -196,6 +196,25 @@ if sys.version_info >= (3, 11):
     SQLITE_WARNING: int
     SQLITE_WARNING_AUTOINDEX: int
 
+if sys.version_info >= (3, 12):
+    LEGACY_TRANSACTION_CONTROL: int
+    SQLITE_DBCONFIG_DEFENSIVE: int
+    SQLITE_DBCONFIG_DQS_DDL: int
+    SQLITE_DBCONFIG_DQS_DML: int
+    SQLITE_DBCONFIG_ENABLE_FKEY: int
+    SQLITE_DBCONFIG_ENABLE_FTS3_TOKENIZER: int
+    SQLITE_DBCONFIG_ENABLE_LOAD_EXTENSION: int
+    SQLITE_DBCONFIG_ENABLE_QPSG: int
+    SQLITE_DBCONFIG_ENABLE_TRIGGER: int
+    SQLITE_DBCONFIG_ENABLE_VIEW: int
+    SQLITE_DBCONFIG_LEGACY_ALTER_TABLE: int
+    SQLITE_DBCONFIG_LEGACY_FILE_FORMAT: int
+    SQLITE_DBCONFIG_NO_CKPT_ON_CLOSE: int
+    SQLITE_DBCONFIG_RESET_DATABASE: int
+    SQLITE_DBCONFIG_TRIGGER_EQP: int
+    SQLITE_DBCONFIG_TRUSTED_SCHEMA: int
+    SQLITE_DBCONFIG_WRITABLE_SCHEMA: int
+
 # Can take or return anything depending on what's in the registry.
 @overload
 def adapt(__obj: Any, __proto: Any) -> Any: ...
@@ -214,8 +233,9 @@ def connect(
 ) -> Connection: ...
 def enable_callback_tracebacks(__enable: bool) -> None: ...
 
-# takes a pos-or-keyword argument because there is a C wrapper
-def enable_shared_cache(enable: int) -> None: ...
+if sys.version_info < (3, 12):
+    # takes a pos-or-keyword argument because there is a C wrapper
+    def enable_shared_cache(enable: int) -> None: ...
 
 if sys.version_info >= (3, 10):
     def register_adapter(__type: type[_T], __adapter: _Adapter[_T]) -> None: ...
@@ -279,6 +299,11 @@ class Connection:
     isolation_level: str | None  # one of '', 'DEFERRED', 'IMMEDIATE' or 'EXCLUSIVE'
     @property
     def total_changes(self) -> int: ...
+    if sys.version_info >= (3, 12):
+        @property
+        def autocommit(self) -> int: ...
+        @autocommit.setter
+        def autocommit(self, val: int) -> None: ...
     row_factory: Any
     text_factory: Any
     def __init__(
@@ -356,6 +381,9 @@ class Connection:
         def getlimit(self, __category: int) -> int: ...
         def serialize(self, *, name: str = "main") -> bytes: ...
         def deserialize(self, __data: ReadableBuffer, *, name: str = "main") -> None: ...
+    if sys.version_info >= (3, 12):
+        def getconfig(self, __op: int) -> bool: ...
+        def setconfig(self, __op: int, __enable: bool = True) -> bool: ...
 
     def __call__(self, __sql: str) -> _Statement: ...
     def __enter__(self) -> Self: ...
