@@ -1,7 +1,7 @@
+from _typeshed import Incomplete
 from collections.abc import ItemsView
-from typing import Any
 
-def ignore_ref_siblings(schema) -> list[tuple[str, Any]] | ItemsView[str, Any]: ...
+def ignore_ref_siblings(schema) -> list[tuple[str, Incomplete]] | ItemsView[str, Incomplete]: ...
 def dependencies_draft3(validator, dependencies, instance, schema) -> None: ...
 def dependencies_draft4_draft6_draft7(validator, dependencies, instance, schema) -> None: ...
 def disallow_draft3(validator, disallow, instance, schema) -> None: ...
