@@ -1,6 +1,5 @@
-from _typeshed import SupportsKeysAndGetItem
+from _typeshed import Incomplete, SupportsKeysAndGetItem
 from collections.abc import Generator, Iterable, Iterator, Mapping, MutableMapping, Sized
-from typing import Any
 
 class URIDict(MutableMapping[str, str]):
     def normalize(self, uri: str) -> str: ...
@@ -16,11 +15,13 @@ class Unset: ...
 
 def load_schema(name): ...
 def format_as_index(container: str, indices) -> str: ...
-def find_additional_properties(instance: Iterable[Any], schema: Mapping[Any, Any]) -> Generator[Any, None, None]: ...
-def extras_msg(extras: Iterable[Any] | Sized) -> str: ...
-def ensure_list(thing) -> list[Any]: ...
+def find_additional_properties(
+    instance: Iterable[Incomplete], schema: Mapping[Incomplete, Incomplete]
+) -> Generator[Incomplete, None, None]: ...
+def extras_msg(extras: Iterable[Incomplete] | Sized) -> str: ...
+def ensure_list(thing) -> list[Incomplete]: ...
 def equal(one, two) -> bool: ...
 def unbool(element, true=..., false=...): ...
 def uniq(container) -> bool: ...
-def find_evaluated_item_indexes_by_schema(validator, instance, schema) -> list[Any]: ...
-def find_evaluated_property_keys_by_schema(validator, instance, schema) -> list[Any]: ...
+def find_evaluated_item_indexes_by_schema(validator, instance, schema) -> list[Incomplete]: ...
+def find_evaluated_property_keys_by_schema(validator, instance, schema) -> list[Incomplete]: ...
