@@ -1,4 +1,4 @@
-from typing import Any
+from _typeshed import Incomplete
 
 class _CannotLoadFile(Exception): ...
 
@@ -25,7 +25,7 @@ class _PlainFormatter:
     def validation_error(self, instance_path, error): ...
     def validation_success(self, instance_path): ...
 
-parser: Any
+parser: Incomplete
 
 def parse_args(args): ...
 def main(args=...) -> None: ...
